**Description**: fix Delete() method #123

**Benchmark before change**:
```
BenchmarkJsonParserLarge-4                    	  200000	     49941 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserMedium-4                   	 1000000	      7363 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserDeleteMedium-4             	 1000000	      6219 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyManualMedium-4      	 1000000	      5056 ns/op	     112 B/op	       2 allocs/op
BenchmarkJsonParserEachKeyStructMedium-4      	 1000000	      5781 ns/op	     672 B/op	      12 allocs/op
BenchmarkJsonParserObjectEachStructMedium-4   	 1000000	      8082 ns/op	     624 B/op	      11 allocs/op
BenchmarkJsonParserSmall-4                    	10000000	       784 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyManualSmall-4       	10000000	       649 ns/op	      80 B/op	       2 allocs/op
BenchmarkJsonParserEachKeyStructSmall-4       	10000000	       744 ns/op	     144 B/op	       4 allocs/op
BenchmarkJsonParserObjectEachStructSmall-4    	10000000	       658 ns/op	     128 B/op	       3 allocs/op
BenchmarkJsonParserSetSmall-4                 	10000000	      1156 ns/op	     816 B/op	       5 allocs/op
BenchmarkJsonParserDelSmall-4                 	 5000000	      1295 ns/op	       0 B/op	       0 allocs/op
```

**Benchmark after change**:
```
BenchmarkJsonParserLarge-4                    	  200000	     50000 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserMedium-4                   	 1000000	      8386 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserDeleteMedium-4             	 1000000	      8880 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyManualMedium-4      	 1000000	      5167 ns/op	     112 B/op	       2 allocs/op
BenchmarkJsonParserEachKeyStructMedium-4      	 1000000	      6114 ns/op	     672 B/op	      12 allocs/op
BenchmarkJsonParserObjectEachStructMedium-4   	 1000000	      8756 ns/op	     624 B/op	      11 allocs/op
BenchmarkJsonParserSmall-4                    	10000000	       841 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyManualSmall-4       	10000000	       648 ns/op	      80 B/op	       2 allocs/op
BenchmarkJsonParserEachKeyStructSmall-4       	10000000	       797 ns/op	     144 B/op	       4 allocs/op
BenchmarkJsonParserObjectEachStructSmall-4    	10000000	       660 ns/op	     128 B/op	       3 allocs/op
BenchmarkJsonParserSetSmall-4                 	 5000000	      1265 ns/op	     816 B/op	       5 allocs/op
BenchmarkJsonParserDelSmall-4                 	 5000000	      1453 ns/op	       0 B/op	       0 allocs/op
```

**Diff benchmarks (benchcmp)**:
```
benchmark                                       old ns/op     new ns/op     delta
BenchmarkJsonParserLarge-4                      49941         50000         +0.12%
BenchmarkJsonParserMedium-4                     7363          8386          +13.89%
BenchmarkJsonParserDeleteMedium-4               6219          8880          +42.79%
BenchmarkJsonParserEachKeyManualMedium-4        5056          5167          +2.20%
BenchmarkJsonParserEachKeyStructMedium-4        5781          6114          +5.76%
BenchmarkJsonParserObjectEachStructMedium-4     8082          8756          +8.34%
BenchmarkJsonParserSmall-4                      784           841           +7.27%
BenchmarkJsonParserEachKeyManualSmall-4         649           648           -0.15%
BenchmarkJsonParserEachKeyStructSmall-4         744           797           +7.12%
BenchmarkJsonParserObjectEachStructSmall-4      658           660           +0.30%
BenchmarkJsonParserSetSmall-4                   1156          1265          +9.43%
BenchmarkJsonParserDelSmall-4                   1295          1453          +12.20%
```